// src/app/models/eventoModel.ts
export interface eventoModel {
    id: number;
    dataHora: Date;
    nomeEvento: string;
    descricaoEvento: string;
    cepEvento: string;
    ruaEvento: string;
    numeroEvento: string;
    bairroEvento: string;
    cidadeEvento: string;
    ufEvento: string;
    complementoEvento: string;
    idResponsavel: number;
}